refactor: Remove TripDialogDesktop and TripDialogMobile
