+++ src/components/Views/Trip.jsx
@@ -2,18 +2,15 @@ import React, { useMemo, useState } from 'react'
 import { useLocation } from 'react-router-dom'
 import SuccessSnackbar from 'src/components/ContactToPlace/SuccessSnackbar'
 import TripProvider from 'src/components/Providers/TripProvider'
-import TripDialogDesktop from 'src/components/Trip/TripDialogDesktop'
-import TripDialogMobile from 'src/components/Trip/TripDialogMobile'
+import TripDialog from 'src/components/Trip/TripDialog'
 import { buildTimeserieQueryById } from 'src/queries/queries'
 
 import { isQueryLoading, useQuery } from 'cozy-client'
 import Spinner from 'cozy-ui/transpiled/react/Spinner'
-import useBreakpoints from 'cozy-ui/transpiled/react/providers/Breakpoints'
 
 const Trip = () => {
   const [ContactToPlaceSuccessMessage, setContactToPlaceSuccessMessage] =
     useState('')
-  const { isMobile } = useBreakpoints()
   const { pathname } = useLocation()
   const timeserieId = useMemo(() => pathname.split('/').pop(), [pathname])
 
@@ -33,11 +30,7 @@ const Trip = () => {
 
   return (
     <TripProvider timeserie={data}>
-      {isMobile ? (
-        <TripDialogMobile onSuccessMessage={setContactToPlaceSuccessMessage} />
-      ) : (
-        <TripDialogDesktop onSuccessMessage={setContactToPlaceSuccessMessage} />
-      )}
+      <TripDialog onSuccessMessage={setContactToPlaceSuccessMessage} />
       <SuccessSnackbar
         message={ContactToPlaceSuccessMessage}
         onClose={() => setContactToPlaceSuccessMessage('')}
